requirements.txt
autopep8==2.0.2
matplotlib==3.7.0
mutagen==1.47.0
numpy==1.26.4
pandas==1.4
Pillow==9.4.0
prompt-toolkit==3.0.38
PyMySQL==1.0.2
python-dotenv==1.0.0
requests==2.28.2
SLPP==1.2.3
tqdm==4.64.1
cryptography
TTS==0.22.0
pydub==0.25.1